requirements.txt
# Core Flask components
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0

# Database components
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
pymysql==1.1.0
cryptography==41.0.4

# Batch processing
schedule==1.2.0

# Deployment
gunicorn==21.2.0
supervisor==4.2.5

# Testing
pytest==8.0.0
pytest-flask==1.3.0 